#596: improve page editor lag
